requirements.txt
eel==0.14.0
requests==2.25.1
python-socketio==5.0.4